переробка PetsData
