+++ src/components/PetsData/PetsData.jsx
@@ -1,51 +1,21 @@
-import PropTypes from "prop-types";
-import { useDeletePetMutation } from "redux/userApi";
-import { ToastContainer, toast } from "react-toastify";
-import "react-toastify/dist/ReactToastify.css";
+import AddPetBtn from "components/AddPetBtn";
 
-import { PetItem, InfoPet, Title, ImgPet, BoxInfo, DeleteBtm, DelIcon } from "./PetsData.styled";
+import { TitlePet, BoxBtn, BoxTitlePet, BoxPet, TitleBtn } from "./PetsData.styled";
+import PetsList from "components/PetsList";
 
-const PetsData = ({ _id, name, birthday, breed, comments, avatar }) => {
-  const [deletePet, { isLoading: isDeleting }] = useDeletePetMutation();
-
-  isDeleting && toast(`${name} was removed`);
-
+const PetsData = () => {
   return (
-    <PetItem>
-      <ImgPet src={`${avatar}`} alt={`${name}`}></ImgPet>
-      <BoxInfo>
-        <DeleteBtm type="button" disabled={isDeleting} onClick={() => deletePet(_id)}>
-          <DelIcon />
-        </DeleteBtm>
-        <InfoPet>
-          <Title>Name: </Title>
-          {name}
-        </InfoPet>
-        <InfoPet>
-          <Title>Date of birth: </Title>
-          {birthday}
-        </InfoPet>
-        <InfoPet>
-          <Title>Breed: </Title>
-          {breed}
-        </InfoPet>
-        <InfoPet>
-          <Title>Comments: </Title>
-          {comments}
-        </InfoPet>
-      </BoxInfo>
-      <ToastContainer />
-    </PetItem>
+    <BoxPet>
+      <BoxTitlePet>
+        <TitlePet>My pets:</TitlePet>
+        <BoxBtn>
+          <TitleBtn>Add pet</TitleBtn>
+          <AddPetBtn disabled={true} />
+        </BoxBtn>
+      </BoxTitlePet>
+      <PetsList />
+    </BoxPet>
   );
 };
 
 export default PetsData;
-
-PetsData.prototype = {
-  _id: PropTypes.string.isRequired,
-  name: PropTypes.string.isRequired,
-  birthday: PropTypes.string.isRequired,
-  breed: PropTypes.string.isRequired,
-  comments: PropTypes.string.isRequired,
-  avatar: PropTypes.string.isRequired,
-};
