+++ src/components/PetsList/PetsList.jsx
@@ -1,14 +1,12 @@
-import { TitlePet, BoxBtn, BoxTitlePet, BoxPet, TitleBtn } from "./PetsList.styled";
+import { useState } from "react";
 import { useGetUserQuery } from "redux/userApi";
 
-import PetsData from "components/PetsData";
-import AddPetBtn from "components/AddPetBtn";
-import { useState } from "react";
+import PetsListItem from "components/PetsListItem";
 
 const PetsList = () => {
   // const pets = [
   //   {
-  //     id: _1,
+  //     id: "_1",
   //     name: "Jack",
   //     data: "22.04.2018",
   //     breed: "JKKKKKK",
@@ -16,7 +14,7 @@ const PetsList = () => {
   //       "Lorem ipsum dolor sit amet consecteturLorem ipsum dolor sit amet, consectetur Lorem ipsum dolor sit amet, consecteturamet consectetur",
   //   },
   //   {
-  //     id: _2,
+  //     id: "_2",
   //     name: "Jack",
   //     data: "22.04.2018",
   //     breed: "JKKKKKK",
@@ -34,23 +32,16 @@ const PetsList = () => {
   }
 
   return (
-    <BoxPet>
-      <BoxTitlePet>
-        <TitlePet>My pets:</TitlePet>
-        <BoxBtn>
-          <TitleBtn>Add pet</TitleBtn>
-          <AddPetBtn disabled={true} />
-        </BoxBtn>
-      </BoxTitlePet>
+    <>
       {isLoading && <div>...Loading</div>}
       {isError ? (
         <div>error</div>
       ) : pets.length === 0 ? (
         <div>Pet not found</div>
       ) : (
-        pets.map(({ _id, name, data, breed, comment }) => <PetsData key={_id} name={name} data={data} breed={breed} comment={comment} id={_id} />)
+        pets.map(({ _id, name, data, breed, comment }) => <PetsListItem key={_id} name={name} data={data} breed={breed} comment={comment} id={_id} />)
       )}
-    </BoxPet>
+    </>
   );
 };
